Changed rating to float and put a limit. Started media rules, done only header atm(it took 2 hours...)

+++ backend/backend/src/main/java/com/dypaworld/model/dto/UserMediaEntryDTO.java
@@ -13,12 +13,12 @@ public class UserMediaEntryDTO {
     Long mediaEntryId;
     String title;
     String imageUrl;
-    int rating;
+    float rating;
     String releaseDate;
     String author;
 
     // Constructor for updating rating
-    public UserMediaEntryDTO(Long mediaEntryId, int rating) {
+    public UserMediaEntryDTO(Long mediaEntryId, float rating) {
         this.mediaEntryId = mediaEntryId;
         this.rating = rating;
     }
